refactor: separate Themes, remove getSubscriptionConfig
